Require earlier version of stripe gem
This is the latest version of the gem that accurately mocks the
version of the stripe API that we are using. Specifically, it allows you
to make calls like `Stripe::Customer.subscriptions.create` whereas
previous versions of the gem do not. Unfortunately, it requires a
version of the Stripe API that is slightly (0.0.1) greater than the one
we are using. This commit simply backs that requirment.